+++ core/main.ts
@@ -23,7 +23,6 @@ export * from "./src/ecs/systems/net/NetServerSystem"
 export * from "./src/graphics/Camera"
 export * from "./src/graphics/Renderer"
 export * from "./src/ecs/systems/ui/ChatBubbleSystem"
-export * from "./src/ecs/systems/ui/ClickableSystem"
 export * from "./src/ecs/systems/ui/HealthBarSystem"
 export * from "./src/ecs/systems/ui/InputSystem"
 export * from "./src/ecs/systems/ui/NametagSystem"
